Migrate to microservices - Testes for User service

+++ user-service/src/main/java/pl/leastsquaresalgorithms/user/repository/UserRepository.java
@@ -4,7 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import pl.leastsquaresalgorithms.user.model.UserEntity;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
 @Repository
@@ -17,5 +16,5 @@ public interface UserRepository extends JpaRepository<UserEntity, Long> {
 
     Optional<UserEntity> findByLoginOrEmail(String login, String email);
 
-    Optional<UserEntity> findByUserIdAndDeleted(BigInteger userId, Boolean deleted);
+    Optional<UserEntity> findByUserIdAndDeleted(Long userId, Boolean deleted);
 }
